Add internal package and move internal classes
